+++ src/fonduer/features/feature_libs/tabular_features.py
@@ -17,7 +17,12 @@ binary_tablelib_feats = {}
 settings = get_config()
 
 
-def get_table_feats(candidates):
+def extract_tabular_features(candidates):
+    """Extract tabular features.
+
+    :param candidates: A list of candidates to extract features from
+    :type candidates: list
+    """
     candidates = candidates if isinstance(candidates, list) else [candidates]
     for candidate in candidates:
         args = tuple([m.context for m in candidate.get_mentions()])
@@ -32,7 +37,7 @@ def get_table_feats(candidates):
             span = args[0]
             if span.stable_id not in unary_tablelib_feats:
                 unary_tablelib_feats[span.stable_id] = set()
-                for f, v in tablelib_unary_features(span):
+                for f, v in _tablelib_unary_features(span):
                     unary_tablelib_feats[span.stable_id].add((f, v))
 
             for f, v in unary_tablelib_feats[span.stable_id]:
@@ -45,7 +50,7 @@ def get_table_feats(candidates):
                 for span, pre in [(span1, "e1_"), (span2, "e2_")]:
                     if span.stable_id not in unary_tablelib_feats:
                         unary_tablelib_feats[span.stable_id] = set()
-                        for f, v in tablelib_unary_features(span):
+                        for f, v in _tablelib_unary_features(span):
                             unary_tablelib_feats[span.stable_id].add((f, v))
 
                     for f, v in unary_tablelib_feats[span.stable_id]:
@@ -53,7 +58,7 @@ def get_table_feats(candidates):
 
                 if candidate.id not in binary_tablelib_feats:
                     binary_tablelib_feats[candidate.id] = set()
-                    for f, v in tablelib_binary_features(span1, span2):
+                    for f, v in _tablelib_binary_features(span1, span2):
                         binary_tablelib_feats[candidate.id].add((f, v))
 
                 for f, v in binary_tablelib_feats[candidate.id]:
@@ -64,17 +69,17 @@ def get_table_feats(candidates):
             )
 
 
-def tablelib_unary_features(span):
+def _tablelib_unary_features(span):
     """
     Table-/structure-related features for a single span
     """
     if not span.sentence.is_tabular():
         return
     sentence = span.sentence
-    for attrib in settings["featurization"]["table"]["unary_features"]["attrib"]:
+    for attrib in settings["featurization"]["tabular"]["unary_features"]["attrib"]:
         for ngram in get_cell_ngrams(
             span,
-            n_max=settings["featurization"]["table"]["unary_features"][
+            n_max=settings["featurization"]["tabular"]["unary_features"][
                 "get_cell_ngrams"
             ]["max"],
             attrib=attrib,
@@ -92,7 +97,7 @@ def tablelib_unary_features(span):
             for ngram in get_head_ngrams(
                 span,
                 axis,
-                n_max=settings["featurization"]["table"]["unary_features"][
+                n_max=settings["featurization"]["tabular"]["unary_features"][
                     "get_head_ngrams"
                 ]["max"],
                 attrib=attrib,
@@ -100,7 +105,7 @@ def tablelib_unary_features(span):
                 yield f"{axis.upper()}_HEAD_{attrib.upper()}_[{ngram}]", DEF_VALUE
         for ngram in get_row_ngrams(
             span,
-            n_max=settings["featurization"]["table"]["unary_features"][
+            n_max=settings["featurization"]["tabular"]["unary_features"][
                 "get_row_ngrams"
             ]["max"],
             attrib=attrib,
@@ -108,7 +113,7 @@ def tablelib_unary_features(span):
             yield f"ROW_{attrib.upper()}_[{ngram}]", DEF_VALUE
         for ngram in get_col_ngrams(
             span,
-            n_max=settings["featurization"]["table"]["unary_features"][
+            n_max=settings["featurization"]["tabular"]["unary_features"][
                 "get_col_ngrams"
             ]["max"],
             attrib=attrib,
@@ -125,11 +130,11 @@ def tablelib_unary_features(span):
         #      yield "COL_INFERRED_%s_[%s]" % (attrib.upper(), ngram), DEF_VALUE
 
 
-def tablelib_binary_features(span1, span2):
+def _tablelib_binary_features(span1, span2):
     """
     Table-/structure-related features for a pair of spans
     """
-    binary_features = settings["featurization"]["table"]["binary_features"]
+    binary_features = settings["featurization"]["tabular"]["binary_features"]
     if span1.sentence.is_tabular() and span2.sentence.is_tabular():
         if span1.sentence.table == span2.sentence.table:
             yield "SAME_TABLE", DEF_VALUE