+++ tests/parser/test_parser.py
@@ -1,4 +1,3 @@
-#! /usr/bin/env python
 import logging
 import os
 from unittest.mock import patch