+++ CHANGELOG.rst
@@ -3,6 +3,35 @@
 
 Added
 ^^^^^
+* `@senwu`_: Refactor `Featurization` to support user defined customized feature
+    extractors and rename existing feature extractors' name to match the paper.
+
+.. note::
+
+    Rather than using a fixed multimodal feature library along, we have added an
+    interface for users to provide customized feature extractors. Please see our
+    full documentation for details.
+
+    .. code:: python
+
+        from fonduer.features import Featurizer, FeatureExtractor
+
+        # Example feature extractor
+        def feat_ext(candidates):
+            for candidate in candidates:
+                yield candidate.id, f"{candidate.id}", 1
+
+        feature_extractors=FeatureExtractor(customize_feature_funcs=[feat_ext])
+        featurizer = Featurizer(session, [PartTemp], feature_extractors=feature_extractors)
+
+    Rather than:
+
+    .. code:: python
+
+        from fonduer.features import Featurizer
+
+        featurizer = Featurizer(session, [PartTemp])
+
 * `@HiromuHota`_: Add page argument to get_pdf_dim in case pages have different dimensions.
 
 Fixed
@@ -38,9 +67,12 @@ Added
 Changed
 ^^^^^^^
 * `@HiromuHota`_: Assign a section for mention spaces.
-* `@HiromuHota`_: Incorporate entity_confusion_matrix as a first-class citizen and rename it to confusion_matrix because it can be used both entity-level and mention-level.
+* `@HiromuHota`_: Incorporate entity_confusion_matrix as a first-class citizen and
+    rename it to confusion_matrix because it can be used both entity-level
+    and mention-level.
 * `@HiromuHota`_: Separate Spacy#_split_sentences_by_char_limit to test itself.
-* `@HiromuHota`_: Refactor the custom sentence_boundary_detector for readability and efficiency.
+* `@HiromuHota`_: Refactor the custom sentence_boundary_detector for readability
+    and efficiency.
 * `@HiromuHota`_: Remove a redundant argument, document, from Spacy#split_sentences.
 * `@HiromuHota`_: Refactor TokenPreservingTokenizer for readability.